Implementing the logic for registering a band and listing the registered bands.
Creating the 'RegistrarBanda' function that captures a band input by the user through the terminal and stores this band in the 'listaDaBandas' variable, which is a collection/list containing the names of the bands provided by the user. Finally, creating the 'MostrarBandasRegistradas' function to display all the bands saved by the user, iterating through each name in the list using the 'foreach' or 'for' loop.