Creating the final feature of the application, the option to display the average of the bands
Implementing the latest feature of the C# application, which is the option to display the average ratings of the registered bands. For this purpose, the 'ExibirMediaDeUmBanda' function was created, which takes input from the user regarding which band they want to see the average ratings for. It then checks if this band is registered and, if it is, provides information about the ratings the band has received and, ultimately, what the rating average for that band is.